In this commit I created 3 screens: Login, Registration and Address. The address screen is a continuation of the registration screen and it was an arbitrary decision to give it this name.
There is a known error on these screens, which is related to the logo, as I haven't yet
decided how to treat them (an image or a string?).

In addition, I have implemented the error screen only for the login so that we have an
idea of how to do it; for the rest, we still need to implement this error alert.

This commit is made to be beaten up, humiliated and cursed at in order to improve the
quality of the screens, so don't be afraid to fire away.

The first idea isn't always the best